tests for price type

+++ dedupe/predicates.py
@@ -269,13 +269,13 @@ def latLongGridPredicate(field, digits=1):
 
 def orderOfMagnitude(field) :
     if field :
-        return (unicode(round(math.log10(field))), )
+        return (unicode(int(round(math.log10(field)))), )
     else :
         return ()
 
 def roundTo1(field) : # thanks http://stackoverflow.com/questions/3410976/how-to-round-a-number-to-significant-figures-in-python
     if field :
-        return (unicode(round(field, -int(math.floor(math.log10(abs(field)))))),)
+        return (unicode(int(round(field, -int(math.floor(math.log10(abs(field))))))),)
     else :
         return ()
